Passing extraFixColorMap to eslint-plugin-uilib settings
