+++ .eslintrc.js
@@ -1,4 +1,15 @@
-const validColors = require("./src/style/colorsPalette")
+const validColors = require("./src/style/colorsPalette");
+const extraColorsDict = {
+  black: 'black',
+  white: 'white',
+  green: 'green30',
+  red: 'red30',
+  blue: 'blue30',
+  '#000': 'black',
+  '#fff': 'white',
+  '#ddd': 'dark70',
+};
+
 
 module.exports = {
   "parser": "babel-eslint",
@@ -38,7 +49,8 @@ module.exports = {
       }
     },
     "uiLib": {
-     "validColors": validColors
+     "validColors": validColors,
+     "extraFixColorMap": extraColorsDict
     }
   }
 }
